Set super dark's menu.selectionBackground
Using a color from elsewhere in the super dark theme to make the currently selected menu item more visible. 

The color used for `activityBar.activeBorder` seems like a good middle ground of not too bright (it is "super dark," after all) and not too dark for the menu (need some contrast).